requirements.txt
numpy
pandas
pyyaml
python-dotenv
ipykernel
openpyxl
matplotlib
seaborn